+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/ProjectSummary.groovy
@@ -2,6 +2,7 @@ package life.qbic.portal.sampletracking.components.projectoverview
 
 import groovy.transform.EqualsAndHashCode
 import java.time.Instant
+import life.qbic.portal.sampletracking.components.projectoverview.statusdisplay.SampleCount
 import life.qbic.business.project.Project
 
 /**
@@ -17,19 +18,19 @@ class ProjectSummary {
     final String code
     final String title
     int totalSampleCount
-    int samplesReceived
-    int samplesQcFailed
-    int sampleDataAvailable
-    int samplesLibraryPrepFinished
+    SampleCount samplesReceived
+    SampleCount samplesQc
+    SampleCount samplesLibraryPrepFinished
+    SampleCount sampleDataAvailable
     Instant lastChanged
 
     ProjectSummary(String code, String title) {
         this.code = code
         this.title = title
-        this.samplesReceived = 0
-        this.samplesQcFailed = 0
-        this.sampleDataAvailable = 0
-        this.samplesLibraryPrepFinished = 0
+        this.samplesReceived = new SampleCount(0,0,0)
+        this.samplesQc = new SampleCount(0,0,0)
+        this.samplesLibraryPrepFinished = new SampleCount(0,0,0)
+        this.sampleDataAvailable = new SampleCount(0,0,0)
         this.totalSampleCount = 0
         this.lastChanged = Instant.MIN
     }